Move binary example to a unit test
