--- example/binary-file/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "binary-file-example",
-  "version": "0.0.0",
-  "description": "",
-  "main": "vite.config.js",
-  "type": "module",
-  "scripts": {
-    "test": "web-test-runner"
-  },
-  "author": "",
-  "license": "UNLICENSED",
-  "private": true,
-  "devDependencies": {
-    "@remcovaes/web-test-runner-vite-plugin": "file:../..",
-    "@open-wc/testing": "^3.1.7",
-    "@testing-library/vue": "^6.6.1",
-    "@types/mocha": "^10.0.1",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "@web/test-runner": "^0.15.0",
-    "@web/test-runner-commands": "^0.6.5",
-    "@web/test-runner-playwright": "^0.9.0",
-    "vite": "^4.1.3"
-  }
-}
