--- example/binary-file/web-test-runner.config.js
@@ -1,8 +1,0 @@
-import { vitePlugin } from '@remcovaes/web-test-runner-vite-plugin';
-
-export default {
-  files: 'src/**/*.test.ts',
-  plugins: [
-    vitePlugin(),
-  ],
-};
