+++ package.json
@@ -22,7 +22,13 @@
   },
   "author": "Remco Vaes",
   "license": "MIT",
+  "scripts": {
+    "test": "web-test-runner && npm run test --ws"
+  },
   "peerDependencies": {
     "vite": "^3.0.0 || ^4.0.0"
+  },
+  "devDependencies": {
+    "@open-wc/testing": "^3.1.7"
   }
 }
