+++ web-test-runner.config.js
@@ -1,0 +1,8 @@
+import { vitePlugin } from './src/index.js';
+
+export default {
+  files: 'test/**/*.test.ts',
+  plugins: [
+    vitePlugin(),
+  ],
+};
